feat: refactor config column proximity maps
