+++ frontend/src/views/DoubleMapView.vue
@@ -18,76 +18,13 @@
     </v-row>
     <v-row class="flex-grow-1" no-gutters>
       <v-col cols="2" height="100%">
-        <v-card flat>
-          <v-card-title>Parameters</v-card-title>
-          <v-card-text>
-            <v-card-title class="text-subtitle-1 px-0 pt-4 pb-0"
-              >Source</v-card-title
-            >
-
-            <v-radio-group hide-details v-model="selectedTilesSource">
-              <v-radio
-                v-for="tilesParams in listTilesParams"
-                class="variable"
-                :key="tilesParams.name"
-                :label="tilesParams.name"
-                :value="tilesParams"
-                density="compact"
-              >
-              </v-radio>
-            </v-radio-group>
-
-            <template
-              v-for="[category, variables] in groupedVariables.entries()"
-            >
-              <v-card-title class="text-subtitle-1 px-0 pt-4 p">
-                {{ cleanVariableString(category) + " weights" }}
-              </v-card-title>
-
-              <template v-for="variable in variables">
-                <v-row align="center" class="my-1" no-gutters>
-                  <v-col class="d-flex" cols="7">
-                    <v-checkbox-btn
-                      v-model="variable.selected"
-                      hide-details
-                      density="compact"
-                      class="text-body-2"
-                    >
-                    </v-checkbox-btn>
-                    <v-label>
-                      {{
-                        cleanVariableString(
-                          getVariableNameWithoutGroup(variable)
-                        )
-                      }}
-                    </v-label>
-                  </v-col>
-
-                  <v-col cols="5">
-                    <v-slider
-                      hide-details
-                      density="compact"
-                      track-size="1"
-                      thumb-size="10"
-                      :color="variable.selected ? 'black' : 'grey'"
-                      min="0"
-                      max="1"
-                      step="0.25"
-                      v-model="variable.weight"
-                      thumb-label
-                    ></v-slider>
-                  </v-col>
-                </v-row>
-              </template>
-            </template>
-          </v-card-text>
-          <v-divider></v-divider>
-          <v-card-actions>
-            <v-btn @click="resetSessionStorage" variant="flat">
-              Reset parameters
-            </v-btn>
-          </v-card-actions>
-        </v-card>
+        <ProximityConfigColumn
+          v-model:variables="variables"
+          :tiles-sources="listTilesParams"
+          v-model:selected-tiles-source="selectedTilesSource"
+          :on-reset-session-storage="resetSessionStorage"
+        >
+        </ProximityConfigColumn>
       </v-col>
 
       <v-divider vertical></v-divider>
@@ -120,16 +57,11 @@
 <script lang="ts" setup>
 import VectorsMap from "@/components/VectorsMap.vue";
 import { ref, computed, watch } from "vue";
-import {
-  listPossibleVariables,
-  listTilesParams,
-  cleanVariableString,
-  getVariableGroup,
-  getVariableNameWithoutGroup,
-} from "@/utils/variables";
+import { listPossibleVariables, listTilesParams } from "@/utils/variables";
 import type { TileParams, ProximityVariable } from "@/utils/variables";
 import type { Map as MaplibreMap } from "maplibre-gl";
 import { syncMaps } from "@/utils/syncmap";
+import ProximityConfigColumn from "@/components/ProximityConfigColumn.vue";
 
 const listVariables = listPossibleVariables;
 
@@ -148,16 +80,6 @@ const storageKeyVariables = "selectedVariables",
     : JSON.parse(JSON.stringify(defaultVariables));
 
 const variables = ref<ProximityVariable[]>(savedVariables);
-
-const groupedVariables = computed(() =>
-  variables.value.reduce((group, variable) => {
-    const category = getVariableGroup(variable);
-    if (group.has(category)) group.get(category).push(variable);
-    else group.set(category, [variable]);
-
-    return group;
-  }, new Map())
-);
 
 const selectedVariables = computed(() => {
   //I added weight so it update props when weight change
@@ -184,9 +106,9 @@ const storageKeyTilesSource = "selectedTilesSource",
 
 const selectedTilesSource = ref(savedTilesSource);
 
-watch(selectedTilesSource, (newTilesSource) =>
-  sessionStorage.setItem(storageKeyTilesSource, JSON.stringify(newTilesSource))
-);
+watch(selectedTilesSource, (newTilesSource) => {
+  sessionStorage.setItem(storageKeyTilesSource, JSON.stringify(newTilesSource));
+});
 
 watch(selectedVariables, () =>
   sessionStorage.setItem(storageKeyVariables, JSON.stringify(variables.value))
